+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 ## [Unreleased]
 - Using lazarus project files 
 
+## [1.1.7] - 2022-12-16
+## add:
+-  add error handle for client to avoid crashing
+-  change maximumCompletions default to 50
+
 ## [1.1.6] - 2022-12-06
 ## add:
 -  add build events